+++ R/mc_extract.r
@@ -18,18 +18,18 @@
 #' data(mcmc)
 #'
 #' # simple extraction
-#' hammer_extract(mcmc, param = 'beta', j = 1:3)
-#' hammer_extract(mcmc, param = 'beta', j = 1:3, stat = 'median')
-#' hammer_extract(mcmc, param = 'z_hat', j = TRUE, k = TRUE)
+#' mc_extract(mcmc, param = 'beta', j = 1:3)
+#' mc_extract(mcmc, param = 'beta', j = 1:3, stat = 'median')
+#' mc_extract(mcmc, param = 'z_hat', j = TRUE, k = TRUE)
 #'
 #' # complex extraction
 #' indices <- list(list(i = TRUE), list(j = TRUE))
 #' params <- c('alpha', 'beta')
-#' hammer_extract(mcmc, param = params, indices = indices)
-#' hammer_extract(mcmc, param = params, indices = indices, stat = 'median')
+#' mc_extract(mcmc, param = params, indices = indices)
+#' mc_extract(mcmc, param = params, indices = indices, stat = 'median')
 #'
-#' @export hammer_extract
-hammer_extract <- function(mcmc, param, i = NULL, j = NULL, k = NULL, l = NULL, indices = NULL, stat = 'mean', quant = 0.5) {
+#' @export mc_extract
+mc_extract <- function(mcmc, param, i = NULL, j = NULL, k = NULL, l = NULL, indices = NULL, stat = 'mean') {
 
 	if (FALSE) {
 
@@ -38,6 +38,8 @@ hammer_extract <- function(mcmc, param, i = NULL, j = NULL, k = NULL, l = NULL, 
 		j <- NULL
 		k <- NULL
 		l <- NULL
+		stat <- 'mean'
+		indices <- NULL
 
 	}
 
@@ -47,9 +49,9 @@ hammer_extract <- function(mcmc, param, i = NULL, j = NULL, k = NULL, l = NULL, 
 	
 		for (count_param in seq_along(param)) {
 		
-			args <- list(mcmc = mcmc, param = param[count_param], stat = stat, quant = quant)
+			args <- list(mcmc = mcmc, param = param[count_param], stat = stat)
 			args <- c(args, indices[[count_param]])
-			this_out <- do.call(hammer_extract, args = args)
+			this_out <- do.call(mc_extract, args = args)
 
 			if (count_param == 1) {
 				out <- this_out
@@ -62,8 +64,8 @@ hammer_extract <- function(mcmc, param, i = NULL, j = NULL, k = NULL, l = NULL, 
 	
 	}
 
-	mcmc_summary <- hammer_summary(mcmc, fail = TRUE)
-	params <- hammer_param(param, i = i, j = j, k = k, l = l, mcmc = mcmc)
+	mcmc_summary <- mc_summary(mcmc, fail = TRUE)
+	params <- mc_param(mcmc = mcmc, param = param, i = i, j = j, k = k, l = l)
 
 	stat <- tolower(stat)
 	stat <- omnibus::pmatchSafe(stat, c('mean', 'median', 'sd', 'lower', 'upper'), nmax = 1)